+++ CHANGES.rst
@@ -5,5 +5,7 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Publish Redis messages on file change
+  [kreafox]
 - Initial release.
   [MrTango]